Website/requirements.txt
Django==2.1.3
requests==2.20.1
django-crispy-forms==1.7.2
django-markdownx
